darts/requirements.txt
Flask==3.0.3
Flask-Mail==0.10.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
scikit-learn==1.5.1
ultralytics==8.2.78
gunicorn==20.1.0
Werkzeug==3.0.3
torch==2.4.0
transformers=4.44.0